pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bsort"
version = "0.1.0"
description = "Bottle cap detection CLI tool for real-time computer vision"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3.10",
]

dependencies = [
    "ultralytics>=8.0.0",
    "opencv-python>=4.8.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "wandb>=0.15.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "pylint>=2.17.0",
    "mypy>=1.5.0",
]

[project.scripts]
bsort = "bsort.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["bsort*"]

[tool.black]
line-length = 88
target-version = ["py310"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.pylint.messages_control]
disable = [
    "C0330",
    "C0326",
    "too-few-public-methods",
    "missing-module-docstring",
]

[tool.pylint.format]
max-line-length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=bsort --cov-report=term-missing"